app/requirements.txt
affinegap==1.12
alembic==1.8.1
attrs==22.1.0
boto3==1.24.81
botocore==1.27.81
BTrees==4.10.1
categorical-distance==1.9
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
datetime-distance==0.1.3
dedupe==2.0.18
dedupe-variable-datetime==0.1.5
dedupe-variable-name==0.0.14
DoubleMetaphone==1.1
et-xmlfile==1.1.0
feedparser==6.0.10
future==0.18.2
haversine==2.7.0
highered==0.2.1
idna==3.4
iniconfig==1.1.1
jmespath==1.0.1
joblib==1.2.0
Levenshtein-search==1.4.5
Mako==1.2.3
MarkupSafe==2.1.1
numpy==1.23.3
openpyxl==3.0.10
packaging==21.3
parseratorvariable==0.0.18
persistent==4.9.1
pluggy==1.0.0
prettytable==3.4.1
probableparsing==0.0.1
probablepeople==0.5.4
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
pydantic==1.10.2
pyhacrf-datamade==0.2.6
PyLBFGS==0.2.0.14
pyparsing==3.0.9
pytest==7.1.3
python-crfsuite==0.9.8
python-dateutil==2.8.2
python-dotenv==0.21.0
requests==2.28.1
rollbar==0.16.3
s3transfer==0.6.0
scikit-learn==1.1.2
scipy==1.9.1
sgmllib3k==1.0.0
simplecosine==1.2
six==1.16.0
SQLAlchemy==1.4.41
sqlalchemy-json==0.5.0
sqlalchemy2-stubs==0.0.2a27
sqlmodel==0.0.8
threadpoolctl==3.1.0
tomli==2.0.1
typing_extensions==4.3.0
Unidecode==1.3.4
urllib3==1.26.12
wcwidth==0.2.5
zope.index==5.2.1
zope.interface==5.4.0
